tests/requirements.txt
black==25.1.0
cryptography==45.0.2
Django==5.2.1
djangorestframework==3.16.0
PyJWT==2.10.1
requests==2.32.3
twine==6.1.0
